gitflow-hotfix-stash: 0.8.7/twitterApiMediaBugFix

+++ imgUploaderLocal.js
@@ -28,7 +28,7 @@ const main = async function () {
         let albumName = albumList[i];
         let albums = imgur.database.findAlbumData({ title: albumName });
         if (albums.length == 0) {
-            // await imgur.api.album.albumCreation({ title: albumName, cover: "vtHXE4B" });
+            // await imgur.api.album.albumCreation({ title: albumName, cover: "Tsupr1Z" });
             console.log("albums.length == 0");
             return;
         }
@@ -128,16 +128,17 @@ const checkImages = async function () {
             // await dbox.fileUpload(`DelImages/${filename}`, body);
 
             await new Promise((resolve, reject) => {
-                require("request").get(img.imageLink)
-                    .pipe(fs.createWriteStream("./" + filename))
-                    .on("error", (e) => { console.log("pipe error", e) })
-                    .on("close", async () => {
-                        let body = fs.readFileSync("./" + filename);
+                require("request").get(img.imageLink, { encoding: 'binary' }, async (error, response, body) => {
+                    if (body) {
+                        fs.writeFileSync("./" + filename, body, { encoding: 'binary' });
+                        body = fs.readFileSync("./" + filename);
                         await dbox.fileUpload("DelImages/" + filename, body);
                         fs.unlinkSync("./" + filename);
                         await imgur.api.image.imageDeletion({ imageHash: img.id });
                         resolve();
-                    });
+                    }
+                    // if (error || !body) { return console.log(error); }
+                });
             });
         }
     }
